server/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpx==0.7.8
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
loguru==0.3.2
outcome==1.2.0
psycopg2-binary==2.9.8
pydantic==0.32.2
pyee==8.2.2
PyJWT==2.8.0
pyppeteer==1.0.2
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tiktok-bot==0.6.4
TikTokApi-Pyppeteer==3.7.9
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==1.26.17
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
social-post-api
